Share hook call visitor logic
