+++ lib/src/analyzer/rules/avoid_hooks_in_control_flow_rule.dart
@@ -2,10 +2,10 @@ import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_context.dart';
 import 'package:analyzer/analysis_rule/rule_visitor_registry.dart';
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/error/error.dart';
 
 import '../utils/utils.dart';
+import 'hook_call_visitor.dart';
 
 class AvoidHooksInControlFlowRule extends AnalysisRule {
   static const LintCode code = LintCode(
@@ -40,81 +40,14 @@ class AvoidHooksInControlFlowRule extends AnalysisRule {
   }
 }
 
-class _AvoidHooksInControlFlowVisitor extends SimpleAstVisitor<void> {
-  final AnalysisRule rule;
-  final bool skip;
-  final CustomHookRegistry customHooks;
-
-  _AvoidHooksInControlFlowVisitor(this.rule, this.skip, this.customHooks);
+class _AvoidHooksInControlFlowVisitor extends HookCallVisitorBase {
+  _AvoidHooksInControlFlowVisitor(super.rule, super.skip, super.customHooks);
 
   @override
-  void visitMethodInvocation(MethodInvocation node) {
-    if (skip) {
-      return;
-    }
-    final hook = matchHookInvocation(node);
-    if (hook != null) {
-      _reportIfNeeded(node, node.methodName, hook.name);
-      return;
-    }
-    if (!customHooks.isCustomHookInvocation(node)) {
-      return;
-    }
-    _reportIfNeeded(node, node.methodName, _hookName(node.methodName));
-  }
-
-  @override
-  void visitFunctionExpressionInvocation(FunctionExpressionInvocation node) {
-    if (skip) {
-      return;
-    }
-    if (!customHooks.isCustomHookInvocationExpression(node)) {
-      return;
-    }
-    final nameNode = customHookInvocationNameNode(node);
-    _reportIfNeeded(node, nameNode, _hookName(nameNode));
-  }
-
-  @override
-  void visitInstanceCreationExpression(InstanceCreationExpression node) {
-    if (skip) {
-      return;
-    }
-    final hook = matchHookConstructor(node);
-    if (hook == null) {
-      return;
-    }
-    _reportIfNeeded(node, node.constructorName, hook.name);
-  }
-
-  void _reportIfNeeded(AstNode node, AstNode target, String hookName) {
-    final scope = enclosingHookScope(node, customHooks: customHooks);
-    if (scope == null) {
-      return;
-    }
+  bool shouldReport(AstNode node, HookScope scope) {
     if (isInsideNestedFunction(node, scope.node)) {
-      return;
+      return false;
     }
-    if (!isInsideControlFlow(node, scope.node)) {
-      return;
-    }
-    final scopeLabel = hookScopeLabel(scope, customHooks: customHooks);
-    rule.reportAtNode(
-      target,
-      arguments: [formatLintArgument(hookName), formatLintArgument(scopeLabel)],
-    );
-  }
-
-  String _hookName(AstNode node) {
-    if (node is SimpleIdentifier) {
-      return node.name;
-    }
-    if (node is PrefixedIdentifier) {
-      return node.identifier.name;
-    }
-    if (node is PropertyAccess) {
-      return node.propertyName.name;
-    }
-    return node.toSource();
+    return isInsideControlFlow(node, scope.node);
   }
 }
